arch(lint): split layer0.go into focused sibling files
layer0.go at 1 203 lines packed seven sub-parsers into one file,
making targeted changes risky and the file hard to navigate.
Split into three focused siblings within the same package:

- layer0_html.go: HTML-block detection (types 1–7), openHTMLBlock,
  htmlBlockCloses, tagName, and related helpers
- layer0_fence.go: fenceInfo, openingFence, closingFence,
  advanceFenceState, tryFence
- layer0_para.go: scanParagraph, markSetextRun, paragraphLeadKind,
  SourceMayHaveCodeBlock, blockquote helpers (blockDepth,
  isLazyContinuation, lineHasNonFenceCode, stripQuoteMarker),
  and list/setext/thematic-break predicates

layer0.go now holds only the exported types, scanner state machine,
and PI/ATX/indented-code sub-parsers (533 lines, down from 1 203).
Pure file reorganisation — no logic changes.
